Big engine update

+++ core/src/com/crazyputting/objects/Ball.java
@@ -3,14 +3,14 @@ package com.crazyputting.objects;
 import com.badlogic.gdx.math.Vector3;
 
 public class Ball {
-    private final double mass = 0.0;
+    private final float mass = 0.0f;
     private final double diameter = 0.0;
     public boolean colliding = false;
     public boolean stopped = false;
     private Vector3 velocity = new Vector3();
     private Vector3 position;
 
-    public double getMass() {
+    public float getMass() {
         return mass;
     }
